requirements-rtx3050.txt
# RTX 3050 Optimized Requirements
# Core AI/ML frameworks
rasa==3.6.13
rasa-sdk==3.6.2
llama-index==0.9.15
chromadb==0.4.18

# Database and storage
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
redis==5.0.1

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Image processing (CPU-optimized for RTX 3050)
opencv-python-headless==4.8.1.78
pillow==10.0.1
easyocr==1.7.0
pytesseract==0.3.10

# NLP and embeddings (lightweight models)
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
celery==5.3.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Memory optimization
psutil==5.9.6
memory-profiler==0.61.0

# API clients
openai==1.3.7
anthropic==0.7.8